requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
pymongo==4.6.0
google-generativeai==0.3.2
pandas==2.3.3
openpyxl==3.1.5
bson==0.5.10
Werkzeug==2.3.7